chore: divide android app from iOS
